[Backend][GPU] Separate GPU lowering pass from LLVM
